pyproject.toml
[tool.poetry]
name = "low-resource-prompting"
version = "0.1.0"
description = ""
authors = ["Asit"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
lightning = "^2.1.0"
torch = "^2.1.0"
hydra-core = "^1.3.2"
transformers = "^4.34.1"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
